+++ geotrek/api/v2/views/sensitivity.py
@@ -7,7 +7,7 @@ from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
 from geotrek.api.v2 import serializers as api_serializers, \
     viewsets as api_viewsets
-from geotrek.api.v2.functions import Transform, Buffer, GeometryType
+from geotrek.api.v2.functions import Transform, Buffer, GeometryType, Area
 from geotrek.sensitivity import models as sensitivity_models
 from ..filters import GeotrekQueryParamsFilter, GeotrekInBBoxFilter, GeotrekSensitiveAreaFilter
 
@@ -46,6 +46,9 @@ class SensitiveAreaViewSet(api_viewsets.GeotrekViewset):
                 When(geom_type='POINT', then=Transform(Buffer(F('geom'), F('species__radius'), 4), settings.API_SRID)),
                 When(geom_type='POLYGON', then=Transform(F('geom'), settings.API_SRID))
             ))
+        # Ensure smaller areas are at the end of the list, ie above bigger areas on the map
+        # to ensure we can select every area in case of overlapping
+        queryset = queryset.annotate(area=Area('geom2d_transformed')).order_by('-area')
         return queryset
 
     def list(self, request, *args, **kwargs):